* Queued Modal Prediction added * Some utilities added * Model updated with new data * Training images moved to GDrive's --BE Project/Coding Stuff/Archived Training Images

+++ GetMyHand/Classification/classifyPythonAPI.hpp
@@ -10,6 +10,6 @@ using namespace std;
 using namespace boost::python;
 
 bool initializePythonInterpreter();
-long predictSignByKNN_Py_Interface(char* data);
+long long predictSignByKNN_Py_Interface(char* data);
 
 #endif
